service-bus-relay/requirements.txt
websockets>=14.0
azure-servicebus>=7.14.0
azure-identity>=1.16.0
hvac>=2.4.0
environs>=15.0.0